Split off RandomSeqGeneratorGlobal.h
